fixed ssl issues
the current build of jruby-openssl 0.6.1 didn't work on appengine--
something to with security violations and the digest lib.

So I reverted to the 0.5.2 release from rc-gems.appspot.com, which works,
but is incompatible with oauth 0.4.0, because its digest bits have a slightly
different set of interface extensions.

you get a ArgumentError - wrong # of arguments(3 for 2):

I fixed this by reverting oauth to 0.3.6 which lines up more neatly.

After all that, it seems to work.